Redefine EncryptSQLException's vendor codes
