+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-core/src/main/java/org/apache/shardingsphere/encrypt/exception/syntax/UnsupportedEncryptInsertValueException.java
@@ -15,18 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.encrypt.exception;
+package org.apache.shardingsphere.encrypt.exception.syntax;
 
+import org.apache.shardingsphere.encrypt.exception.EncryptSQLException;
 import org.apache.shardingsphere.infra.util.exception.external.sql.sqlstate.XOpenSQLState;
 
 /**
- * Unsupported encrypt SQL exception.
+ * Unsupported encrypt insert value exception.
  */
-public final class UnsupportedEncryptSQLException extends EncryptSQLException {
+public final class UnsupportedEncryptInsertValueException extends EncryptSQLException {
     
-    private static final long serialVersionUID = 8391663664617661217L;
+    private static final long serialVersionUID = 5004882561157380582L;
     
-    public UnsupportedEncryptSQLException(final String sqlClause) {
-        super(XOpenSQLState.SYNTAX_ERROR, 21, "The SQL clause `%s` is unsupported in encrypt rule", sqlClause);
+    public UnsupportedEncryptInsertValueException(final int columnIndex) {
+        super(XOpenSQLState.SYNTAX_ERROR, 20, "Insert value of index `%s` can not support for encrypt", String.valueOf(columnIndex));
     }
 }
