+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-core/src/main/java/org/apache/shardingsphere/encrypt/exception/syntax/UnsupportedEncryptSQLException.java
@@ -15,18 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.encrypt.exception;
+package org.apache.shardingsphere.encrypt.exception.syntax;
 
+import org.apache.shardingsphere.encrypt.exception.EncryptSQLException;
 import org.apache.shardingsphere.infra.util.exception.external.sql.sqlstate.XOpenSQLState;
 
 /**
- * Encrypt column not found exception.
+ * Unsupported encrypt SQL exception.
  */
-public final class EncryptColumnNotFoundException extends EncryptSQLException {
+public final class UnsupportedEncryptSQLException extends EncryptSQLException {
     
-    private static final long serialVersionUID = -8481487069938556675L;
+    private static final long serialVersionUID = 8391663664617661217L;
     
-    public EncryptColumnNotFoundException(final String tableName, final String columnName) {
-        super(XOpenSQLState.CHECK_OPTION_VIOLATION, 1, "Fail to find encrypt column `%s` from table `%s`", columnName, tableName);
+    public UnsupportedEncryptSQLException(final String sqlClause) {
+        super(XOpenSQLState.SYNTAX_ERROR, 21, "The SQL clause `%s` is unsupported in encrypt rule", sqlClause);
     }
 }
