+++ shardingsphere-features/shardingsphere-shadow/shardingsphere-shadow-core/src/main/java/org/apache/shardingsphere/shadow/algorithm/shadow/validator/ShadowValueValidator.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.shadow.algorithm.shadow.validator;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.apache.shardingsphere.shadow.exception.UnsupportedShadowColumnTypeException;
+import org.apache.shardingsphere.shadow.exception.data.UnsupportedShadowColumnTypeException;
 
 import java.util.Date;
 import java.util.HashSet;
