+++ shardingsphere-features/shardingsphere-shadow/shardingsphere-shadow-core/src/test/java/org/apache/shardingsphere/shadow/algorithm/shadow/validator/ShadowValueValidatorTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.shadow.algorithm.shadow.validator;
 
-import org.apache.shardingsphere.shadow.exception.UnsupportedShadowColumnTypeException;
+import org.apache.shardingsphere.shadow.exception.data.UnsupportedShadowColumnTypeException;
 import org.junit.Test;
 
 import java.util.Date;
